[RHELC-937] Port check_dbus_is_running to Action framework
