+++ convert2rhel/actions/__init__.py
@@ -315,7 +315,6 @@ def perform_system_checks():
     check_custom_repos_are_valid()
     check_tainted_kmods()
     is_loaded_kernel_latest()
-    check_dbus_is_running()
 
 
 def perform_pre_ponr_checks():
@@ -764,21 +763,3 @@ def is_loaded_kernel_latest():
         )
 
     logger.info("The currently loaded kernel is at the latest version.")
-
-
-def check_dbus_is_running():
-    """Error out if we need to register with rhsm and the dbus daemon is not running."""
-    logger.task("Prepare: Check that DBus Daemon is running")
-
-    if tool_opts.no_rhsm:
-        logger.info("Skipping the check because we have been asked not to subscribe this system to RHSM.")
-        return
-
-    if system_info.dbus_running:
-        logger.info("DBus Daemon is running")
-        return
-
-    logger.critical(
-        "Could not find a running DBus Daemon which is needed to register with subscription manager.\n"
-        "Please start dbus using `systemctl start dbus`"
-    )
