+++ convert2rhel/unit_tests/actions/actions_test.py
@@ -157,7 +157,6 @@ def test_perform_pre_checks(monkeypatch):
     check_thirdparty_kmods_mock = mock.Mock()
     check_custom_repos_are_valid_mock = mock.Mock()
     is_loaded_kernel_latest_mock = mock.Mock()
-    check_dbus_is_running_mock = mock.Mock()
 
     monkeypatch.setattr(
         actions,
@@ -175,13 +174,10 @@ def test_perform_pre_checks(monkeypatch):
         value=check_custom_repos_are_valid_mock,
     )
     monkeypatch.setattr(actions, "is_loaded_kernel_latest", value=is_loaded_kernel_latest_mock)
-    monkeypatch.setattr(actions, "check_dbus_is_running", value=check_dbus_is_running_mock)
-
     actions.perform_system_checks()
 
     check_thirdparty_kmods_mock.assert_called_once()
     is_loaded_kernel_latest_mock.assert_called_once()
-    check_dbus_is_running_mock.assert_called_once()
 
 
 def test_perform_pre_ponr_checks(monkeypatch):
@@ -1006,43 +1002,3 @@ class TestIsLoadedKernelLatest:
         uname_kernel_version = uname_version.rsplit(".", 1)[0]
         assert "Latest kernel version available in baseos: %s" % repoquery_kernel_version in caplog.records[-1].message
         assert "Loaded kernel version: %s\n\n" % uname_kernel_version in caplog.records[-1].message
-
-
-@pytest.mark.parametrize(
-    ("no_rhsm", "dbus_running", "log_msg"),
-    (
-        (True, True, "Skipping the check because we have been asked not to subscribe this system to RHSM."),
-        (True, False, "Skipping the check because we have been asked not to subscribe this system to RHSM."),
-        (False, True, "DBus Daemon is running"),
-    ),
-)
-def test_check_dbus_is_running(
-    caplog, monkeypatch, global_tool_opts, global_system_info, no_rhsm, dbus_running, log_msg
-):
-    monkeypatch.setattr(actions, "tool_opts", global_tool_opts)
-    global_tool_opts.no_rhsm = no_rhsm
-    monkeypatch.setattr(actions, "system_info", global_system_info)
-    global_system_info.dbus_running = dbus_running
-
-    assert actions.check_dbus_is_running() is None
-    assert caplog.records[-1].message == log_msg
-
-    assert log_msg == caplog.records[-1].message
-
-
-def test_check_dbus_is_running_not_running(caplog, monkeypatch, global_tool_opts, global_system_info):
-    monkeypatch.setattr(actions, "tool_opts", global_tool_opts)
-    global_tool_opts.no_rhsm = False
-    monkeypatch.setattr(actions, "system_info", global_system_info)
-    global_system_info.dbus_running = False
-
-    with pytest.raises(SystemExit):
-        actions.check_dbus_is_running()
-
-    log_msg = (
-        "Could not find a running DBus Daemon which is needed to"
-        " register with subscription manager.\nPlease start dbus using `systemctl"
-        " start dbus`"
-    )
-    assert log_msg == caplog.records[-1].message
-    assert caplog.records[-1].levelname == "CRITICAL"
